% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_3.R
\name{price_volatility}
\alias{price_volatility}
\title{Calculate Price Volatility for a Locality}
\usage{
price_volatility(property_data, Locality)
}
\arguments{
\item{property_data}{Data frame containing property data with columns 'Locality' and 'median_price'.}

\item{Locality}{Character string specifying the name of the locality for which to measure volatility.}
}
\value{
A list with 'Locality' and 'Price Volatility', representing the standard deviation of prices for the locality.
}
\description{
Measures the volatility of property prices over the available period for a specified locality.
}
\examples{
price_volatility(property_data, "ABBOTSFORD")
}
